Make HTML Beautifier fail on nesting errors
This way we're also checking mistakes like closing tags that don't match
their opening element, which we detected by manually running HTML
Beautifier with the `-e` option, and fixed two commits ago.

Note there was a false positive in the mailer layout. Apparently,
closing the ERB tag right before the HTML opening tag and the lack of
other attributes on the tag made HTML Beautifier think the tag wasn't
correctly open. We're fixing it by adding an HTML id attribute to the
element.