Move tabs link list outside tabs content
Just like we do everywhere else. We're also removing the wrong ARIA
attributes that we added in commit c18479e3a, which caused an
accessibility issue reported by Axe:

```
1) aria-required-children: Certain ARIA roles must contain particular
   children (critical)
   https://dequeuniversity.com/rules/axe/4.10/aria-required-children?application=axeAPI
   The following 1 node violate this rule:

     Selector: .tabs-content
     HTML: <div class="tabs-content" data-tabs-content="information-texts-tabs" role="tablist">
     Fix any of the following:
     - Element has children which are not allowed: ul[tabindex]
```

Although, in this case, it would probably be better to have different
pages instead of tabs, so loading the page doesn't take too long.